feat: extract db engine creation logic in core
