+++ scripts/playbooks.py
@@ -20,7 +20,7 @@ from pathlib import Path
 import click
 import networkx as nx
 
-from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params import collections_option
 from tdp.core.constants import DEFAULT_SERVICE_PRIORITY, SERVICE_PRIORITY
 from tdp.core.dag import Dag
 from tdp.core.entities.operation import OperationName
