+++ tdp/cli/params.py
@@ -131,7 +131,7 @@ def database_dsn_option(func: FC) -> FC:
     Return a SQLAlchemy Engine instance, available as "db_engine" in the command context.
     """
 
-    from sqlalchemy import create_engine
+    from tdp.core.db import get_engine
 
     return click.option(
         "db_engine",
@@ -139,7 +139,7 @@ def database_dsn_option(func: FC) -> FC:
         envvar="TDP_DATABASE_DSN",
         required=True,
         type=str,
-        callback=lambda _ctx, _param, value: create_engine(value, future=True),
+        callback=lambda _ctx, _param, value: get_engine(value),
         help=(
             "Database Data Source Name, in sqlalchemy driver form "
             "example: sqlite:////data/tdp.db or sqlite+pysqlite:////data/tdp.db. "
